Add deprecation warning for older k8s versions
As discussed during yesterday's office hours.

This should be cherry-picked back to 1.16.
We'll need to bump the minor version in the 1.17 and master branches.
Once we officially drop support we could make this a sliding window that uses the `kops.Version` variable.
I think we could have deprecation warnings for the 2 oldest versions still supported by the kops version, announcing that the n+2 Kops version will no longer support the specified k8s version.