OpenStack: Use interface attach when reconciling server ports
This makes use of the interface attach method when reconciling server
ports.

The difference between just setting the `DeviceID` on the Port and this
is that with the attachment process the server is also validated. Which
means for example that a Port cannot be attached to a server in `ERROR`
state.